feat: publish scheduler durability and HITL observability hardening

+++ pkg/graph/state_store_integration_test.go
@@ -54,7 +54,7 @@ func TestFileStateStore_AgentRoundTrip(t *testing.T) {
 	}
 
 	// Ensure version carried over.
-	if restored.Version() != snap.Version {
+	if restored.Version() != uint64(snap.Version) {
 		t.Fatalf("version mismatch: got %d want %d", restored.Version(), snap.Version)
 	}
